run.py MAJOR DATA REVISION MOVE TO HOME ACTION
Move to home is now an action during data collection.
This ensures that all the poses of the objects are stored successfully, and some data is collected after the stack is complete. It should now also be possible to apply corrections to the dataset if failures and successes are mislabeled, and error failures such as path planning errors or security stops are now also saved.